flesh out codegen backend trait
This commit adds a skeleton codegen backend by implementing the
rustc CodegenBackend trait. This trait informs rustc of valid
backends through use of the unstable `codegen-backend` flag
available in the nightly compiler.

This commit also adds a very simple example crate (in the examples
folder) that implements the simplest valid `no_std` program for
testing the backend as it evolves. A utility function called
`compile_bell` will compile the backend and then attempt to compile
the example using the generated backend. It is available in the dev
shell provided by the nix file. Refer to the README to see how to
drop into a development shell using nix.