fix rust-analyzer issues on rustc_* crates
This commit fixes an issue caused by running rust-analyzer on
private rustc source, which causes cargo to attempt to create a
lockfile in the read-only nix store. This is a well-known problem
in the Rust community without any solutions, so this commit works
around the issue by appending a manually generated lockfile (on
`aarch64-darwin` using the toolchain specified in the included
toolchain config) to the correct place in the rust toolchain
source.

Relevant GitHub issues below:
- https://github.com/rust-lang/cargo/issues/10096
- https://github.com/rust-lang/rust-analyzer/issues/13393
- https://github.com/rust-lang/rust/issues/95736