autogenerate rustc-src Cargo.lock
This commit removes the statically defined Cargo.lock in favor of
autogenerating it as needed during the nix env preparation process
as a build dependency.

The current solution requires recompiling the toolchain to include
this fix, but we should look into using `symlinkJoin` to allow for
composing the toolchain out of prebuilt binaries for faster setup.

The main issue is that using `symlinkJoin` requires wrapping rustc
and exposing a bunch of environmental variables to inform rustc of
the new symlink sysroot that includes the autogenerated Cargo.lock