add scaffolding for CodegenCx and Builder traits (#21)
* add scaffolding for CodegenCx and Builder traits

This commit adds the scaffolding necessary for the QirCodegenBackend to
be considered as a valid codegen backend for `rustc_codegen_ssa`, the
generic backend for custom compiler plugins.

This required the addition of two new structs: `QirCodegenCompiler` and
`QirBuilder`, each with their own set of scaffolded traits. Methods
which have yet to be implemented have been added to an `unknown.rs`
file to make it easy to spot missing impls as we flesh out each struct.